@@ -18,7 +18,7 @@ description: "A new Flutter module project."
 version: 1.0.0+1
 
 environment:
-  sdk: '3.2.4'
+  sdk: '3.3.4'
 
 dependencies:
   flutter:
@@ -31,11 +31,27 @@ dependencies:
   syncfusion_flutter_charts: ^25.1.40
   syncfusion_flutter_gauges: ^25.1.40
   flutter_overlay_window: any
+  call_log: ^4.0.0
+  flutter_background_service: ^4.0.0
+  flutter_local_notifications: ^14.1.1
+  flutter_background_service_android: ^5.0.0
+  phone_state_background: ^0.0.4
+  shared_preferences: ^2.2.3
+  http: ^1.2.0
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+  path_provider: ^2.1.3
+  jwt_decode: ^0.3.1
+  contacts_service: ^0.6.3
+  permission_handler: ^11.3.1
+  phone_state: ^1.0.4
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.0
+  hive_generator: ^2.0.1
+  build_runner: ^2.4.9
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -84,7 +100,7 @@ flutter:
   # maintain consistency when adding or modifying assets and plugins.
   # They also do not have any bearing on your native host application's
   # identifiers, which may be completely independent or the same as these.
-  module:
-    androidX: true
-    androidPackage: com.example.secucalls
-    iosBundleIdentifier: com.example.secucalls
+  # module:
+  #   androidX: true
+  #   androidPackage: com.example.secucalls
+  #   iosBundleIdentifier: com.example.secucalls
