@@ -1,18 +1,33 @@
-
-
 // ignore_for_file: constant_identifier_names
 
 const String title_appbar = "Tìm lại mật khẩu";
-const String text_title = "Nhập số điện thoại để tìm lại mật khẩu ";
-const String hint_text_phone = "Số điện thoại";
+const String text_title = "Nhập email để tìm lại mật khẩu";
+const String hint_text_email = "Email";
 const String text_button = "Gửi";
 const String text_register_button = "Đăng ký";
 
+// OTP Validation
+const String text_title_otp_validation =
+    "Xác nhận OTP được gửi tới email nếu email bạn nhập chính xác";
+const double top_margin_form_otp = 50;
+const String hint_text_otp = "OTP";
+
+// New Password
+const String text_title_new_password = "Nhập lại mật khẩu mới";
+const double top_margin_form_new_password = 74;
+const String hint_text_new_password = "Mật khẩu mới";
+const double top_margin_register_button_new_password = 44;
+const String text_token_expired =
+    "OTP Token đã hết hiệu lực. OTP mới đã được gửi. Vui lòng nhập lại OTP";
+const String text_change_password_success =
+    "Đổi mật khẩu thành công. Vui lòng đăng nhập lại";
+const String text_otp_validate_success = "OTP hợp lệ";
+
 const double top_margin_title_text = 84;
-const double left_margin_title_text = 48;
+const double left_margin_title_text = 108;
 const double top_margin_form = 103;
 const double left_margin_form = 55;
-const double space_between_form_button = 30;
+const double space_between_form_button = 10;
 
 const double top_margin_register_button = 104;
-const double left_margin_register_button = 290;
+const double left_margin_register_button = 270;
