@@ -3,9 +3,9 @@
 
 const double top_margin_logo = 35;
 const double left_margin_logo = 124;
-const double top_margin_form = 130;
+const double top_margin_form = 80;
 const double left_margin_form = 35;
-const double space_between_text_field = 30;
+const double space_between_text_field = 10;
 
 const double top_margin_two_buttons = 20;
 
@@ -17,3 +17,6 @@ const String text_password = "Mật khẩu";
 const String text_button = "Đăng nhập";
 const String text_register_button = "Đăng ký";
 const String text_forget_password = "Quên mật khẩu";
+const String text_login_fail_snack_bar = "Đăng nhập thất bại";
+const String text_mail_not_exist_snack_bar = "Email không tồn tại";
+const String text_mail_wrong_password_snack_bar = "Mật khẩu sai";
