@@ -1,15 +1,16 @@
 
 // ignore_for_file: constant_identifier_names
 
-const double top_margin_form = 50;
+const double top_margin_form = 20;
 const double left_margin_form = 55;
-const double space_between_text_fields = 30;
-const double top_margin_forget_password_button = 190;
-const double left_margin_forget_password_button = 210;
+const double space_between_text_fields = 10;
+const double top_margin_forget_password_button = 30;
+const double left_margin_forget_password_button = 190;
 
 const String title_appbar = "Đăng ký tài khoản";
 const String hint_text_last_name = "Họ";
 const String hint_text_first_name = "Tên";
+const String hint_text_email = "Email";
 const String hint_text_phone = "Số điện thoại";
 const String hint_text_password = "Mật khẩu";  
 const String text_button = "Đăng ký";
